Replace elastic tests
